pip_requirements.txt
numpy
pandas
sklearn
scipy
pygam
statsmodels
shap
matplotlib
pywavelets
scipy
pathos
census
multiprocess
PyWavelets
PyYAML